Close #32: Simplify configuration of data folder
Instead of allowing a fine tuned configuration where to store the user
and group data, we simply store these in the content/ folder, where it
is supposed to be already well protected. We still offer to have a
single set of members for the whole website, or a separate set for each
language, though.