Don't store derivable information in the session
Firstly, storing the `session_id()` in the session is useless at best,
and evil at worst, at that defies regenerating the session ID to avoid
session fixation. Not storing the `session_id()` fixes #37.

Secondly, the user record is already stored in the user file, and we
don't want to keep both stores in sync.