Fix #46: add a member in backend w/o password doesn't work
If an empty password is submitted (what usually only can happen when a
new user is added in the plugin administration), we create a random
password.  Since the password is not know to anybody, afterwards the
user has to go through the password forgotten procedure.  This allows
webmasters to add users without ever knowing their passwords.