Close #43: Access protection via page data tab
It's pleasant to remove this complex regex for detecting `access()`
calls on the pages.  We rename the function right away, but keep
`access()` as alias, to not break the access protection for users who
have not moved the manual calls to the page data tab, for now.

We need to apply some additional styling to the page data tab, to avoid
a totally borked layout; `XH_helpIcon()` should actually render as
`<span>` not as floating `<div>`.