Fix #59: Registration and password forgotten allow for email probing
For password forgotten, we just don't send the mail, if there is no
user with that mail, and state that we only sent the mail if it exists
in the success message.

For registration, we remind the user in the mail that there is already
an account, and refer to password forgotten, so they can get details
about it.  For that purpose, we always enable the password forgotten
functionality, but keep the option to not display the link in the login
form.