Refactor DB::lock()
Instead of doing locking and unlocking in a single method (storing the
lock file handle in a static variable), we split that in two methods,
where `::lock()` now returns the handle (or `null` what we ignore like
before), while `::unlock()` expects that handle.  Besides getting rid
of an ugly static variable, the client is no longer required to keep
the strict sequence of lock/unlock, and is now aware of what is going
on.