Refactor ForgotPasswordController to action
To help with that we extract minimalist Request and Url classes.  And
for simplicity, we also drop the base_64url encoding; hexits are good
enough.

Hmm, don't like some code duplication, though.  Either something is
generally wrong with our action concept – do we need a superclass,
trait or more delegation?  We'll see.