Make View::message() like ::text()
We should not pass some arbitrary message to `View::message()`, but
rather just the language key, like we do for '::text()`.  This requires
us to use some proper i18n, instead of composing messages by string
concatenation, what cannot work, generally.

We also introduce `View::messagep()` which creates a pluralized message.

A nice side-effect of this modification is that we no longer need to
pass the language array to a couple of controllers/actions.