Proper i18n for err_colon (second attempt)
Of course, it doesn't work if we just use another language string due
to grammar issues.  For instance, in German we would get "Vollständiger
Benutzername darf keinen Doppelpunkt enthalten!"  This is not proper
German.  Since err_colon is only used for the "full name", we can write
a proper message in the first place.