Validator should be language agnostic
Instead of returning a list of error messages, we return a list of list
of tuples consisting of the language key, and possibly further details.
This can be rendered to a proper error message in an outer layer.

To make it possible to step-wise refactor to this, we introduce a new
Validator class, which will eventually replace the ValidationService.