Bring back grouping (kinda)
Previously, error messages have been grouped by user; this was useful
in case of bulk updates with many errors for different users.  While we
don't want to bring back the nested lists, we show the message about
user errors as info, giving some visual separation.  And after all,
contrary to 979e980, this message is not really an error message.