Fix #65 (Add new member - password not hashed)
If the password doesn't match the old password, the password has been
changed by the admin, and we need to hash it regardless of errors.

Now that we hash even on errors, we have a problem with one of the
approval tests; either we wrap `password_hash()` so that we can use a
test double, or we drop the test.  We decide for the latter, but add a
respective `ProcessUsers` test, what might be even better.