Streamline call of request handlers
As is, calling a request handler requires a hard to read (and error-
prone) incantation of creating the handler, calling it, and finally to
`fire()` the response.  Maybe even worse, the names of the handlers are
totally unclear: do they specify the handler object or its callable
behavior?

We introduce `Dic::__callStatic()` to automate all that for us, and
thereby clarifying the name of the handlers: they should be `handle*`.
We postpone the actual renaming for a while, though.

Furthermore, we automatically inject the request into the handlers, as
most of the time, the other parameters are the parameters of the user
functions we are forwarding.  They could be variadic, in which case the
request at the end would be a problem.

* Automatically pass $request to handlers
* Request should always be the first parameter of handlers
* Put magic incantation of Dic::make* behind the scenes
* Make Dic::makeCurrentUser() private
* Dic::makeUserRepository() can be private