Revise CurrentUser
We start the session early (if the session cookie is set), and read the
current username from the session.  If needed, we fetch the full `User`
from the repository.  This trims down the contrived `CurrentUser` to a
very simple `LoginManager` which is just a facade over the required
session functionality.

* Exclude untestable top-level scripts from coverage report
* Change CurrentUser to simple LoginManager
* Replace most uses of CurrentUser with Request and UserRepository
* Start session early (we misuse required_classes.php)