Fix group administration submission
This was broken by 9520006, due to completely wrong assumptions about
what actually is submitted.  There never was the possibility that "add"
and "delete" were set at the same time, and while "delete" was an array
it only ever contained at most one element.  Besides fixing the issue,
we clarify by introducing "action" which can now be used to clearly
distinguish which action is triggered.

Still, the whole concept is not a proper HTTP implementation; "add" and
"delete" should be GET requests (since the stored data are not changed
by these actions), but we leave that as is for now.  In the long run we
are likely better off by introducing a proper CRUD user interface,
anyway.