Replace XH\CSRFProtection with our own CsrfProtector
That class is badly designed, since it starts a session in the c'tor,
it needs an explicit call to `::store()` (except for the admin), and
probably worst of all, `::check()` is too easy to miss since it is a
void function.  Fortunately, implementing a CSRF protector is only a
couple of lines.