Introduce Groups model
Groups maintains the `UserGroup`s in the `DocumentStore`, so there is
no more need for the `UserGroupRepository`.

Accordingly, we turn `UserGroup` into an entity instead of a value
object.

We also do no longer create groups.csv right away, what implies that
the default user group is no longer stored upfront – that was a minor
convenience at best, anyway.